+++ type/carousel/classes/content_type.php
@@ -65,6 +65,7 @@ class content_type extends \mod_unilabel\content_type {
      * @return void
      */
     public function add_form_fragment(\mod_unilabel\edit_content_form $form, \context $context) {
+        global $OUTPUT;
 
         $unilabeltyperecord = $this->load_unilabeltype_record($form->unilabel->id);
 
@@ -103,6 +104,14 @@ class content_type extends \mod_unilabel\content_type {
         $mform->addElement('advcheckbox', $prefix.'usemobile', get_string('use_mobile_images', 'unilabeltype_carousel'));
         $mform->addHelpButton($prefix.'usemobile', 'use_mobile_images', 'unilabeltype_carousel');
 
+        // Prepare the activity url picker.
+        $formid = $mform->getAttribute('id');
+        $course = $form->get_course();
+        $picker = new \mod_unilabel\output\component\activity_picker($course, $formid);
+        $inputidbase = 'id_'.$prefix.'url_';
+        $pickerbutton = new \mod_unilabel\output\component\activity_picker_button($formid, $inputidbase);
+        $mform->addElement('html', $OUTPUT->render($picker));
+
         $repeatarray = [];
         // If we want each repeated elment in a numbered group we add a header with '{no}' in its label.
         // This is replaced by the number of element.
@@ -111,12 +120,22 @@ class content_type extends \mod_unilabel\content_type {
                                 'editor',
                                 $prefix.'caption',
                                 get_string('caption', 'unilabeltype_carousel').'-{no}',
-                                array('rows' => 4));
+                                array('rows' => 4)
+        );
         $repeatarray[] = $mform->createElement(
                                 'text',
                                 $prefix.'url',
                                 get_string('url', 'unilabeltype_carousel').'-{no}',
-                                array('size' => 50));
+                                array('size' => 50)
+        );
+        $repeatarray[] = $mform->createElement(
+            'static',
+            $prefix.'activitypickerbutton',
+            '',
+            $OUTPUT->render($pickerbutton)
+
+        );
+
         $repeatarray[] = $mform->createElement(
             'filemanager',
             $prefix.'image',